fix: handle websocket url properly and mixed security protocols properly
If server behind proxy is using HTTP but frontside of proxy is using HTTPS pivot to wss...

First check if a custom WebSocket URL is provided in the `window.webssh2Config.socket.url`.

If a custom URL is provided, it updates the protocol to wss: or ws: based on whether the current page is served over HTTPS or HTTP.

If no custom URL is provided, it constructs a WebSocket URL using the current location's hostname and automatically selects the appropriate protocol and port.